requirements.txt
PyMuPDF
streamlit
google-generativeai
faiss-cpu
sentence-transformers